refactor: :recycle: use new error classes
