+++ src/shared/errors/unhandledError.ts
@@ -1,10 +1,11 @@
 import { resolveLoggerProvider } from '@shared/container/providers/logger';
-import AppError from './AppError';
+import { HTTP_STATUS } from '@shared/constants/httpStatus';
+import { ErrorResponse } from './error.dtos';
 
 const logger = resolveLoggerProvider();
 
-export const unhandledError = (error: unknown) => {
+export const unhandledError = (error: unknown): ErrorResponse => {
 	const err = error as Error;
 	logger.error(err, `UnhandledError(${err.name}): ${err.message}`);
-	return new AppError('Internal Server Error', 500);
+	return { statusCode: HTTP_STATUS.internal_server_error, message: 'Internal Server Error' };
 };
