+++ src/shared/infra/http/middlewares/ensureAuthenticated.ts
@@ -1,15 +1,15 @@
 import auth from '@config/auth';
-import AppError from '@shared/errors/AppError';
 import { AUTHENTICATION_ERRORS } from '@modules/accounts/errors/authentication.errors';
 import { USER_ERRORS } from '@modules/accounts/errors/user.errors';
 import { UsersRepository } from '@modules/accounts/infra/mongo/repositories/Users.repository';
 import { JsonwebtokenProvider } from '@shared/container/providers/jwt/implementations/Jsonwebtoken.provider';
 import { AppNextFunction, AppRequest, AppResponse } from '../types';
+import { UnauthorizedError } from '@shared/errors/error.lib';
 
 export async function ensureAuthenticated(request: AppRequest, response: AppResponse, next: AppNextFunction) {
 	const authHeader = request.headers.authorization;
 	if (!authHeader) {
-		throw new AppError(AUTHENTICATION_ERRORS.missing_token, 401);
+		throw new UnauthorizedError(AUTHENTICATION_ERRORS.missing_token, { path: 'ensureAuthenticated.middleware' });
 	}
 
 	const [, token] = authHeader.split(' ');
@@ -23,7 +23,7 @@ export async function ensureAuthenticated(request: AppRequest, response: AppResp
 
 		const user = await usersRepository.findById(user_id);
 		if (!user) {
-			throw new AppError(USER_ERRORS.not_exist, 401);
+			throw new UnauthorizedError(USER_ERRORS.not_exist, { path: 'ensureAuthenticated.middleware' });
 		}
 
 		request.user = {
@@ -32,6 +32,6 @@ export async function ensureAuthenticated(request: AppRequest, response: AppResp
 
 		next();
 	} catch {
-		throw new AppError(AUTHENTICATION_ERRORS.invalid_token, 401);
+		throw new UnauthorizedError(AUTHENTICATION_ERRORS.invalid_token, { path: 'ensureAuthenticated.middleware' });
 	}
 }
