+++ src/shared/infra/http/middlewares/errorHandler.ts
@@ -1,5 +1,5 @@
 import { P, match } from 'ts-pattern';
-import AppError from '@shared/errors/AppError';
+import { AppError } from '@shared/errors/error.lib';
 import { AppNextFunction, AppRequest, AppResponse } from '@shared/infra/http/types';
 import {
 	handleMongoError,
@@ -9,9 +9,10 @@ import {
 } from '@shared/infra/mongo/mongo.errors';
 import { handleAppError } from '@shared/errors/handleAppError';
 import { unhandledError } from '@shared/errors/unhandledError';
+import { ErrorResponse } from '@shared/errors/error.dtos';
 
 export async function errorHandler(err: Error, _request: AppRequest, response: AppResponse, _next: AppNextFunction) {
-	const { statusCode, message } = match(err)
+	const { statusCode, message }: ErrorResponse = match(err)
 		.with(P.instanceOf(AppError), handleAppError)
 		.when(instanceOfMongoError, handleMongoError)
 		.when(instanceOfMongooseError, handleMongooseError)
